+++ src/Data/ApiClasses/Base/BaseApiClass.cs
@@ -220,11 +220,10 @@ namespace TheGamesDBApiWrapper.Data.ApiClasses.Base
 
                 if (restResponse.StatusCode == System.Net.HttpStatusCode.OK)
                 {
-                    if (restResponseBase == null)
+                    if (restResponseBase != null)
                     {
-                        throw new Exception("Error: Response deserialization failed. BaseApiResponseModel is null.");
+                        this.allowanceTracker.SetAllowance(restResponseBase.RemainingMonthlyAllowance, restResponseBase.ExtraAllowance, restResponseBase.AllowanceRefreshTimer);
                     }
-                    this.allowanceTracker.SetAllowance(restResponseBase.RemainingMonthlyAllowance, restResponseBase.ExtraAllowance, restResponseBase.AllowanceRefreshTimer);
                     return response;
                 } 
                 else if (restResponse.StatusCode == System.Net.HttpStatusCode.Forbidden && restResponseBase != null)