+++ src/Models/Responses/Countries/CountriesResponse.cs
@@ -1,0 +1,10 @@
+namespace TheGamesDBApiWrapper.Models.Responses.Countries
+{
+    /// <summary>
+    /// CountriesResponse - response for fetching all countries
+    /// </summary>
+    /// <seealso cref="TheGamesDBApiWrapper.Models.Responses.Base.BaseApiResponseModel{TheGamesDBApiWrapper.Models.Responses.Countries.CountriesDataModel}" />
+    public class CountriesResponse : Base.BaseApiResponseModel<CountriesDataModel>
+    {
+    }
+}